for save

+++ zst_project/render.py
@@ -20,7 +20,7 @@ class ApiRenderer(BaseRenderer):
             if response.status_code == 200:
                 response_dict['data'] = data
             else:
-                response_dict['code'] = int(response.status_code / 100 * 1000 + response.status_code % 100)
+                response_dict['code'] = int(int(response.status_code / 100) * 1000 + response.status_code % 100)
                 response_dict['message'] = data
                 response.status_code = 200
